Fix(upgrade): Field 'app_key' doesn't have a default value QUERY : INSERT INTO remote_servers (#504)
Refs: MON-15744

+++ centreon/www/install/php/Update-22.04.8.php
@@ -39,8 +39,10 @@ try {
     $errorMessage = 'Unable to update illegal characters fields from engine configuration of pollers';
     decodeIllegalCharactersNagios($pearDB);
 
-    $errorMessage = "Unable to remove app_key column";
-    $pearDB->query("ALTER TABLE `remote_servers` DROP COLUMN `app_key`");
+    if ($pearDB->isColumnExist('remote_servers', 'app_key') === 1) {
+        $errorMessage = "Unable to remove app_key column";
+        $pearDB->query("ALTER TABLE `remote_servers` DROP COLUMN `app_key`");
+    }
 
     $pearDB->commit();
 } catch (\Exception $e) {
